move time to standalone file
